currently the handmaiden logic needs some fixes
the code was running infinitely in some games. it was because sometimes all other players were untargetable and the choice selection would work out to "while True". fixed it for the RandomAI player. for testing, in runs.py, chnage it to just the two random players for now. itll keep happening with the other AI.

changed the logic in the LoveLetterAI to properly raise an exception when it happens (indexerror) instead of running forever.
added a little ugly logic check before a couple cards for the "None" target type.
also the rulebook I found off of google says if this is happening where everyone's handmaiden'd up and you have the prince, you have to self-target. so I encoded that into the conditional for the randomai. Would probably need ot add that to the other two AI if we want to use heuristics.